fix & refactor: move firebaseConfig to single file
