+++ webview-ui/src/context/FirebaseAuthContext.tsx
@@ -1,21 +1,7 @@
-import { User, getAuth, signInWithCustomToken, signOut } from "firebase/auth"
-import { initializeApp } from "firebase/app"
+import { User, signInWithCustomToken, signOut } from "firebase/auth"
 import React, { createContext, useCallback, useContext, useEffect, useMemo, useState } from "react"
+import { auth } from "../firebaseConfig"
 import { vscode } from "@/utils/vscode"
-
-// Firebase configuration from extension
-const firebaseConfig = {
-	apiKey: "AIzaSyDcXAaanNgR2_T0dq2oOl5XyKPksYHppVo",
-	authDomain: "cline-bot.firebaseapp.com",
-	projectId: "cline-bot",
-	storageBucket: "cline-bot.firebasestorage.app",
-	messagingSenderId: "364369702101",
-	appId: "1:364369702101:web:0013885dcf20b43799c65c",
-	measurementId: "G-MDPRELSCD1",
-}
-// Initialize Firebase
-const app = initializeApp(firebaseConfig)
-const auth = getAuth(app)
 
 interface FirebaseAuthContextType {
 	user: User | null
